Implement model schema parser
The schema parser uses the go/loader package to compile all files in a
Go package (including their direct and indirect dependencies). Then, it
analyzes the parsed ASTs and the resolved type information to construct
the set of models (primary or embedded) that are defined within the
package.

Model fields can drive the schema parser via schema annotations (tags). A
schema annotation has the following format:

`schema:"backendFieldName,attr..."`

backendFieldName specifies the preferred name for this field when it is
persisted to the generated backend stores. If not specified, a
(lower) camel-cased version from the struct field name will be used instead.
The following attributes are supported:
- "pk". Specifies that the field holds the primary key for the model.
   Currently, only a single PK field is supported.
- "find_by". Specifies that this field can be used to find instances
  (plural) of this model; generally such queries yield back iterators to
  the result set.

Any field that is a pointer, a map or a slice is flagged as a nullable
field.

The parser's output (Model, and Field values) will eventually be used to
drive the generator's templates.