Add template for generating model structs
The template generates a struct for each non-embedded model where fields
are private and getters/setters are provided for manipulating their
values. In addition to the model fields, the generated structs contain
two internal fields:
- _version: A int64 that tracks the current model version in the store.
  It is also passed as an argument to the model c-tor
- _flags: A uint8 mask that tracks whether the model instance is new,
  deleted, read from or written to. This information will be used at a
  later stage by the store to select the appropriate operations for
  persisting models.

In addition, a struct for each embedded model will also be defined.
However, the fields of these structs are public to match the semantics
of cross-package types that may also be used as fields for non-embedded
model (e.g. foo map[string]otherpkg.Value).