Add template for generating inmemory backend and test suite
The generated in-memory implementation maintains data in maps. Each
model UUID is assigned its own dedidacted collection of maps for each
model type. The implementation is naive in that it will perform full
table scans for all Find$ModelBy$Field queries. This can be optimized in
the future by implementing aux one-to-n index maps for the each of the
fields that we can search by.